+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@the-devoyage/graphql-users",
-  "version": "0.2.0",
+  "version": "0.2.1",
   "description": "",
   "main": "dist/server.js",
   "scripts": {
